+++ lib/llm/provider.rb
@@ -324,6 +324,7 @@ class LLM::Provider
   # @return [Boolean]
   def streamable?(stream)
     stream.respond_to?(:on_content) ||
+      stream.respond_to?(:on_tool_call) ||
       stream.respond_to?(:on_reasoning_content) ||
       stream.respond_to?(:<<)
   end